Refactors and improves cohesion metrics calculation
Introduces a dedicated service for calculating cohesion metrics,
improving code organization and testability.

Simplifies and corrects the calculation of LCOM metrics, ensuring
accurate cohesion assessment.

Updates LCOM2 to return an integer, aligning with its definition.

+++ architecture_evaluator_backend/src/main/java/io/github/opsord/architecture_evaluator_backend/modules/parser/processor/dto/parts/CohesionMetricsDTO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class CohesionMetricsDTO {
     private Integer lackOfCohesion1;
-    private Double lackOfCohesion2;
+    private Integer lackOfCohesion2;
     private Integer lackOfCohesion3;
     private Integer lackOfCohesion4;
     private Double lackOfCohesion5;
